Remove oj.file
